+++ graphics/java/android/graphics/PorterDuffXfermode.java
@@ -32,5 +32,5 @@ public class PorterDuffXfermode extends Xfermode {
         native_instance = nativeCreateXfermode(mode.nativeInt);
     }
     
-    private static native int nativeCreateXfermode(int mode);
+    private static native long nativeCreateXfermode(int mode);
 }
